+++ CHANGELOG.md
@@ -1,13 +1,13 @@
 # Changelog
 
-## UNRELEASED
+## 2.41.0
 
 Fixed:
   - `mappersmith`: Allow path to be empty string in manifest #327
   - `mappersmith`: `Response.errors` should only be allowed to contain `Error` or `string` #325
 
 Added:
-  - `mappersmith`: Accept `path` as a resource method param
+  - `mappersmith`: Accept `path` as a resource method param #328
 
 ## 2.40.0